+++ fishbase/fish_date.py
@@ -305,3 +305,64 @@ def transform_datetime_to_unix(dtime=None):
         raise TypeError('dtime should be datetime, but we got {}'.format(type(dtime)))
 
     return time.mktime(dtime.timetuple())
+
+
+# v1.1.3 edit by Hu Jun #116
+class FishDateTimeFormat(object):
+    """
+    实现 datetime 和 str 之间相互转换，基于 Python 的 datetime.datetime 进行封装和扩展；
+
+    举例如下::
+
+        print('--- FishDateTimeFormat demo ---')
+        datetime_obj = date(year=2018, month=11, day=23)
+        print(FishDateTimeFormat.strftime(datetime_obj, '%Y-%m-%d'))
+        date_time_str = '2018-11-23 23:17:20'
+        time_format = '%Y-%m-%d %H:%M:%S'
+        print(FishDateTimeFormat.strptime(date_time_str, time_format))
+        print('---')
+
+    执行结果::
+
+        --- FishDateTimeFormat demo ---
+        2018-11-23
+        2018-11-23 23:17:20
+        <class 'datetime.datetime'>
+        ---
+
+        """
+    @staticmethod
+    def strftime(date_time=None, time_format=None):
+        """
+        将 datetime 对象转换为 str
+
+        :param:
+            * date_time: (obj) datetime 对象
+            * time_format: (sting) 日期格式字符串
+        :return:
+            * date_time_str: (string) 日期字符串
+        """
+        if not date_time:
+            datetime_now = datetime.now()
+        else:
+            datetime_now = date_time
+        if not time_format:
+            time_format = '%Y/%m/%d %H:%M:%S'
+        return datetime.strftime(datetime_now, time_format)
+
+    @staticmethod
+    def strptime(time_str, time_format):
+        """
+        将 str 转换为 datetime 对象
+
+        :param:
+            * time_str: (string) 日期字符串
+            * time_format: (sting) 日期格式字符串
+        :return:
+            * datetime_obj: (obj) datetime 对象
+        """
+        try:
+            datetime_obj = datetime.strptime(time_str, time_format)
+            return datetime_obj
+        except ValueError as ex:
+            raise ValueError(ex)